fix: Bogus and very messy hack to dll-interface related warnings

+++ Source/CTF/ctf/Collections/Hash.cpp
@@ -22,20 +22,20 @@
 
 namespace CTF::Collections {
 
-template <typename T> int Hash<T>::operator()(const T &key) const {
-  const unsigned char *data = reinterpret_cast<const unsigned char *>(&key);
-  int Hash = 0;
-  for (int i = 0; i < sizeof(T); ++i)
-    Hash = Hash * 131 + data[i];
-  return Hash;
-}
+    template <typename T> int Hash<T>::operator()(const T &key) const {
+        const unsigned char *data = reinterpret_cast<const unsigned char *>(&key);
+        int Hash = 0;
+        for (int i = 0; i < sizeof(T); ++i)
+            Hash = Hash * 131 + data[i];
+        return Hash;
+    }
 
-int Hash<String>::operator()(const String &s) const {
-  int Hash = 5381;
-  const char *str = s.CStr();
-  int len = s.Length();
-  for (int i = 0; i < len; ++i)
-    Hash = ((Hash << 5) + Hash) + static_cast<unsigned char>(str[i]);
-  return Hash;
-}
+    int Hash<String>::operator()(const String &s) const {
+        int Hash = 5381;
+        const char *str = s.CStr();
+        size_t len = s.Length();
+        for (int i = 0; i < len; ++i)
+            Hash = ((Hash << 5) + Hash) + static_cast<unsigned char>(str[i]);
+        return Hash;
+    }
 } // namespace CTF::Collections
